1613 - cast id_per_locationid_import_date to unsigned int 32 in filter function

+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -1706,14 +1706,14 @@ class EstimateFilledPostsByJobRoleCleanUtilsSchemas:
     }
 
     test_job_role_group_equal_zero_schema = {
-        IndCQC.id_per_locationid_import_date: pl.Int64,
+        IndCQC.id_per_locationid_import_date: pl.UInt32,
         IndCQC.main_job_role_clean_labelled: pl.Enum(
             AscwdsWorkerValueLabelsJobGroup.all_roles()
         ),
         IndCQC.ascwds_job_role_counts: pl.Int64,
     }
     expected_job_role_group_equal_zero_schema = {
-        IndCQC.id_per_locationid_import_date: pl.Int64,
+        IndCQC.id_per_locationid_import_date: pl.UInt32,
         IndCQC.main_job_role_clean_labelled: pl.Enum(
             AscwdsWorkerValueLabelsJobGroup.all_roles()
         ),
